rename files in folder components
