+++ client/src/components/LibraryCatalogList/LibraryCatalogList.scss
@@ -1,11 +1,41 @@
-.Library__Catalog-List {
-  // border: 1px solid black;
-  padding: 20px;
-  text-align: center;
+.Library__Catalog-Item {
+  border: 1px solid black;
+  padding: 30px;
+  border-radius: 20px;
+  margin-bottom: 40px;
 }
 
-.Library__Catalog-List-Zero {
-  text-align: center;
-  margin-top: 30px;
-  font-size: 28px;
+.Library__Catalog-Item-Title {
+  font-size: 22px;
+  font-weight: 500;
+  padding-bottom: 40px;
+}
+
+.Library__Catalog-Item-Descr {
+  font-size: 18px;
+  font-weight: 300;
+  padding-bottom: 40px;
+}
+
+.Library__Catalog-Item-Author {
+  font-size: 18px;
+  font-weight: 300;
+  padding-bottom: 10px;
+  padding-top: 10px;
+}
+
+.Library__Button-Item {
+  button {
+    width: 50%;
+    background-color: #d0d0e5;
+    color: #fff;
+    &::before {
+      background-color: #babad3;
+    }
+  }
+}
+
+.Library__Catalog-Item-Author-Title {
+  font-weight: 400;
+  text-transform: uppercase;
 }
