+++ client/src/pages/MyBooksPage/MyBooks.jsx
@@ -2,7 +2,7 @@ import { useSelector } from "react-redux";
 import "./MyBooks.scss";
 import { useNavigate } from "react-router-dom";
 import { STATIC } from "../../utils/const";
-import MyBooksList from "../../components/MyBooks/MyBooksList";
+import MyBooksList from "../../components/MyBooksList/MyBooksList";
 import { useCallback } from "react";
 import { useAction } from "../../store";
 import Button from "../../components/UI/Button/Button";
